eventbridge-sns-cdk/requirements.txt
aws_cdk.aws_sns
aws_cdk.aws_sns_subscriptions
aws_cdk.aws_events
aws_cdk.aws_events_targets
aws-cdk.core
